check to see if last_warner is valid
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@4111 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/Eiffel/switch/interface/graphical/windows.e
@@ -63,7 +63,7 @@ feature {NONE}
 			!! Result.make (new_parent);
 			Result.set_window (new_parent);
 			old_warner := last_warner;
-			if old_warner /= Void then
+			if old_warner /= Void and then not old_warner.destroyed then
 				old_warner.popdown;
 				old_warner.destroy
 			end;
